change dir of xpu moe weight only
